Account for dupe IO requests from retries
If we're downloading N bytes from a GetObjectTask,
and we make it up to (and queue) bytes N/2 before hitting
a network error, we'll retry the entire request.  This means
that bytes 0-N/2 will be sent to the `DeferQueue` again.

This handles that case by ignoring any requests for offsets
we've already seen.